app/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manoa_agent"
version = "0.1.0"
requires-python = "==3.10.*"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb==0.4.13",
    "datasets>=3.6.0",
    "html2text>=2025.4.15",
    "joblib>=1.5.0",
    "langchain",
    "langchain-chroma>=0.2.3",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.4",
    "langchain-google-vertexai>=2.0.23",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.16",
    "langgraph==0.2.40",
    "langserve>=0.3.1",
    "lxml==5.2.1",
    "neo4j>=5.28.1",
    "neo4j-graphrag>=1.7.0",
    "notebook>=7.4.2",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "rapidfuzz>=3.13.0",
    "scikit-learn==1.5.1",
    "spacy>=3.8.5",
    "sse-starlette>=2.3.5",
]

[project.scripts]
start-hoku = "manoa_agent.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.3.5",
]
